+++ discord/ext/commands/help.py
@@ -186,7 +186,7 @@ def _not_overriden(f):
 
 class _HelpCommandImpl(Command):
     def __init__(self, inject, *args, **kwargs):
-        super().__init__(inject.command_callback, *args, **kwargs)
+        super().__init__(inject.command_callback, *args, add_slash_command=False, **kwargs)
         self._original = inject
         self._injected = inject